Move out styled table tags to module
